feat: impl account data stack
