+++ crates/miden-lib/src/errors/tx_kernel_errors.rs
@@ -135,6 +135,12 @@ pub const ERR_ACCOUNT_STORAGE_SLOT_INDEX_OUT_OF_BOUNDS: u32 = 0x20152;
 pub const ERR_ACCOUNT_TOO_MANY_PROCEDURES: u32 = 0x20153;
 /// Number of account storage slots exceeds the maximum limit of 255
 pub const ERR_ACCOUNT_TOO_MANY_STORAGE_SLOTS: u32 = 0x20154;
+/// Depth of the nested FPI calls exceeded 64
+pub const ERR_ACCOUNT_DATA_STACK_OVERFLOW: u32 = 0x20155;
+/// Attempt to end foreign context, but we already in native context
+pub const ERR_ACCOUNT_DATA_STACK_UNDERFLOW: u32 = 0x20156;
+/// Creation of the foreign context using native account is forbidden
+pub const ERR_ACCOUNT_FOREIGN_CONTEXT_WITH_NATIVE_ACCOUNT: u32 = 0x20157;
 
 /// Provided foreign account ID is equal to the native account ID.
 pub const ERR_FOREIGN_ACCOUNT_ID_EQUALS_NATIVE_ACCT_ID: u32 = 0x20180;
@@ -207,7 +213,7 @@ pub const ERR_VAULT_NON_FUNGIBLE_ASSET_TO_REMOVE_NOT_FOUND: u32 = 0x20286;
 /// Failed to remove fungible asset from the asset vault due to the initial value being invalid
 pub const ERR_VAULT_REMOVE_FUNGIBLE_ASSET_FAILED_INITIAL_VALUE_INVALID: u32 = 0x20287;
 
-pub const TX_KERNEL_ERRORS: [(u32, &str); 86] = [
+pub const TX_KERNEL_ERRORS: [(u32, &str); 89] = [
     (ERR_KERNEL_PROCEDURE_OFFSET_OUT_OF_BOUNDS, "Provided kernel procedure offset is out of bounds"),
 
     (ERR_PROLOGUE_EXISTING_ACCOUNT_MUST_HAVE_NON_ZERO_NONCE, "Existing accounts must have a non-zero nonce"),
@@ -266,6 +272,9 @@ pub const TX_KERNEL_ERRORS: [(u32, &str); 86] = [
     (ERR_ACCOUNT_STORAGE_SLOT_INDEX_OUT_OF_BOUNDS, "Provided storage slot index is out of bounds"),
     (ERR_ACCOUNT_TOO_MANY_PROCEDURES, "Number of account procedures exceeds the maximum limit of 256"),
     (ERR_ACCOUNT_TOO_MANY_STORAGE_SLOTS, "Number of account storage slots exceeds the maximum limit of 255"),
+    (ERR_ACCOUNT_DATA_STACK_OVERFLOW, "Depth of the nested FPI calls exceeded 64"),
+    (ERR_ACCOUNT_DATA_STACK_UNDERFLOW, "Attempt to end foreign context, but we already in native context"),
+    (ERR_ACCOUNT_FOREIGN_CONTEXT_WITH_NATIVE_ACCOUNT, "Creation of the foreign context using native account is forbidden"),
 
     (ERR_FOREIGN_ACCOUNT_ID_EQUALS_NATIVE_ACCT_ID, "Provided foreign account ID is equal to the native account ID."),
     (ERR_FOREIGN_ACCOUNT_ID_IS_ZERO, "ID of the provided foreign account equals zero."),
