+++ crates/miden-lib/src/transaction/procedures/kernel_v0.rs
@@ -10,47 +10,47 @@ pub const KERNEL0_PROCEDURES: [Digest; 36] = [
     // account_get_initial_commitment
     digest!("0x920898348bacd6d98a399301eb308478fd32b32eab019a5a6ef7a6b44abb61f6"),
     // account_get_current_commitment
-    digest!("0x69f2e0b681498846c049d298ae9833649061eeca6552363396d470a49c3e9890"),
+    digest!("0x1c294b4d92eedce11252353164d712200adfa5cc443ac3667dd86ef33b303c8c"),
     // account_get_id
-    digest!("0x0fe7d10d95c529eaec3f0672dd716105c8c5d0ab52e63adaa5c176844bb94a38"),
+    digest!("0x1a5583b3a4011d0ca83ac9633fc12b0c6ec2cba03ee8c5f380ac69fc4f767075"),
     // account_get_nonce
-    digest!("0x0db82eda6e834383e79ee83751434aabadbf71996ce8897562a77a34b43f9a31"),
+    digest!("0xf1dfe3621b9147803b6668352915be7fb7f85df476c9d18052272270a854fa75"),
     // account_incr_nonce
-    digest!("0xe33735d9c594a0515ee30348a4e988a10b941792a4823c69435a37310ec1fd5f"),
+    digest!("0x726357558767a56aae315c4c8aa406cdf57ad400832222622eec58dc42cce406"),
     // account_get_code_commitment
-    digest!("0xf998788832427ab137eaee1c4d99aaea435754b861fcb1d00b1640345eb9afe7"),
+    digest!("0xbab83830e881bdbee08fa1506d651388c20ebb0cadfb6794189542dd257841aa"),
     // account_get_storage_commitment
-    digest!("0x723581bfeac05bc1f233083c80379c1cd049d0c77c0476fab9d0a896aaa76476"),
+    digest!("0xe4cd056f358b2438c3b378a31650ddbbf8d631f135bf966e01a2b57eebd458a5"),
     // account_get_item
-    digest!("0x05b08e5241a702b90b8623cd97ba83f7c608941dcc73ad4f2ffc71a7f04eb61e"),
+    digest!("0xdd8f439cb6f7f3edcda15c9c339e7c2b2dada2fc94952a8199081a197aeebb7a"),
     // account_set_item
-    digest!("0x1d87539d6557cc970fc78bfde945d867858c2e6183bee1dc88aceb9b0acfbe21"),
+    digest!("0x61104ec016c3ed9b49aee53650ddde9e984a72e4c4e13001cbf98b9cef426758"),
     // account_get_map_item
-    digest!("0xc00e1cd25d6a412624d8f69ec74630688868a56e1c7ad94125ef3eb0da80acd6"),
+    digest!("0x21237825d10004f77d3e7b32aee9052b519752fd03f839c19440e2010f73132e"),
     // account_set_map_item
-    digest!("0x8391236934b8353ffc13c6ea6fc61db30d1ac65406fafa0a1ba9153ef3d35c6c"),
+    digest!("0x876168292d11aea0e2886ee6d7f9e723a95ec3aa4e467e6f46c898567bfc3604"),
     // account_get_vault_root
-    digest!("0x814bb6069c9b62fb40597f49547c9ae88603dd9a28f86e9ef0e54e8fc094a92e"),
+    digest!("0x279b4a9e5adca07f01cadf8ecc1303fa3c670003a7a4e69f09506b070c4023df"),
     // account_add_asset
-    digest!("0xbdebb57af4389e79dcf94529225e0e94053958814624f9731a29d984426e39d0"),
+    digest!("0x7ae966cc7da065c7d06248ccdc5cf319eddba55c063900fc97833f39870234f1"),
     // account_remove_asset
-    digest!("0x9d006820999b7e30c6cfc1ffceac01ac5afbbccd1d8ddf787170b9854734135f"),
+    digest!("0x50935b368b7843258ae86f6392df9346721b60dd9563d77a2d548da8ab81e44c"),
     // account_get_balance
-    digest!("0x75add0a0be7be690bb744a904408de0c5293b928f573f981280a96607b2b88ed"),
+    digest!("0xc3385953bc66def5211f53a3c44de8facfb4060abbb1c9708859c314268989e8"),
     // account_has_non_fungible_asset
-    digest!("0x653ab7a20e5af62c0022850726fef4f5fd6468f9de4cfc43b8fb6b9ff12e6b32"),
+    digest!("0x4fea67ed25474d5494a23c5e1e06a93f8aa140d0a673c6e140e0d4f1dd8bd835"),
     // faucet_mint_asset
-    digest!("0x2b755f5f43f1f8957225c7d4c2130a37a4ba15c1323703efe82d3ed676051535"),
+    digest!("0x836ecba4b7caa770b24ffa51115e2c08724ecb7f0854ee3a94848bec5b0ee2ce"),
     // faucet_burn_asset
-    digest!("0xdfa2005118b0bcefe94082b25b09bb3b47f525a026dffad60664d9bb679ca140"),
+    digest!("0x37e6b7c80f478c11ea7f12c065013c4bc3a09100fd6ecc4893e17d0abd675254"),
     // faucet_get_total_fungible_asset_issuance
-    digest!("0x7c46ed8cc84a0c5439285f715d1c867eb71131e9f0b1bbd65acea9dddc35bd96"),
+    digest!("0xd2ee4bd330f989165ee2be0f121a4db916f95e58f6fd2d040d57672f2f0cef63"),
     // faucet_is_non_fungible_asset_issued
-    digest!("0xab87249b2584dc957876f0f2bfc697ad6a9a4d6710e751dc51272f1fbecce5b4"),
+    digest!("0x04c49c1f9cc628c1447e19d46c307084e995044faa86154353eec6af6b5b5041"),
     // note_get_assets_info
     digest!("0x34e4f1ea83eb4342ab8f5acec89962b2ab4b56d9c631e807d8e4dc8efd270bf2"),
     // note_add_asset
-    digest!("0x2785bb643703ca8215f646c0e6a356f7c8eff9991d9ca0aae5d3b87d0a63ffad"),
+    digest!("0x999a790c639966f85b4420f7b00a5965db163e7a69fa288b059f79162f3f81a8"),
     // note_get_serial_number
     digest!("0x59b3ea650232049bb333867841012c3694bd557fa199cd65655c0006edccc3ab"),
     // note_get_inputs_commitment
@@ -60,7 +60,7 @@ pub const KERNEL0_PROCEDURES: [Digest; 36] = [
     // note_get_script_root
     digest!("0x66fb188ca538d9f8bc6fd1aedbd19336bf6e3a1c0ae67b5f725cbc9cb4f7867f"),
     // tx_create_note
-    digest!("0x3c8757885c515be9429d44113766faa5ee9036162fbf238c557615c37c53aa30"),
+    digest!("0xb0ce6d0bd9dfa8b77408fac615eb570381d28769914e7ca0475ba5dd00a2283a"),
     // tx_get_input_notes_commitment
     digest!("0x16cb840dc9131e2fd2b3e83b8d796eb466722ae36f29f27b4b053f1bee2ed473"),
     // tx_get_output_notes_commitment
@@ -72,11 +72,11 @@ pub const KERNEL0_PROCEDURES: [Digest; 36] = [
     // tx_get_block_timestamp
     digest!("0x786863e6dbcd5026619afd3831b7dcbf824cda54950b0e0724ebf9d9370ec723"),
     // tx_start_foreign_context
-    digest!("0xa22ff1895f7e5b266b2d67378e198ed059940a509894d5a43e07689224ac1de9"),
+    digest!("0x56be2bed7d7c9a21a6c26383e586ecb0d65afed2b662ad7ab7246162e5c01fa3"),
     // tx_end_foreign_context
-    digest!("0x0a11755ed547d42974a83aeb7e1df9408aec0bb7423afdd8afdf2a50ef346832"),
+    digest!("0xc7f8665baf9fa71191860633ad3b58cc2b39e83e89342b26e36124920d2403ec"),
     // tx_get_expiration_delta
-    digest!("0x4c3ca7bb7dac8ae20aefe6ebe582499730cd5ffa3d2592ac88b83a4d72873089"),
+    digest!("0x756352beed1624a42d4540c434a4faa986d6d9d08ef8437699d9086fcd9ad9e7"),
     // tx_update_expiration_block_num
-    digest!("0xa7b1045569f0905558f38454bfc4b6bbbd07648e34248161a4bb44cfb557043d"),
+    digest!("0x11ca0c8662d20e6b05fbff4a20423bfa52595862b6c7c5c5ef1cc0a917e4cb62"),
 ];
