+++ crates/miden-tx/src/tests/kernel_tests/mod.rs
@@ -15,6 +15,7 @@ mod test_asset;
 mod test_asset_vault;
 mod test_epilogue;
 mod test_faucet;
+mod test_fpi;
 mod test_note;
 mod test_prologue;
 mod test_tx;
